Fix #3: Extract InFlightMessage from ClientEntityManagerImpl
-forthcoming changes add to the InFlightMessage and add unit tests so splitting the extraction from the actual changes will make the intent of the change clearer
