+++ dso-l1/src/main/java/com/tc/object/ClientEntityManagerImpl.java
@@ -46,8 +46,6 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.Semaphore;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicLong;
 
 
@@ -407,127 +405,5 @@ public class ClientEntityManagerImpl implements ClientEntityManager {
     NetworkVoltronEntityMessage message = (NetworkVoltronEntityMessage) channel.createMessage(TCMessageType.VOLTRON_ENTITY_MESSAGE);
     message.setContents(clientID, transactionID, entityDescriptor, type, requiresReplication, config, oldestTransactionPending);
     return message;
-  }
-
-
-  /**
-   * This is essentially a wrapper over an in-flight VoltronEntityMessage, used for tracking its response.
-   * The message is stored here, since it is sent asynchronously, along with storage for the return value.
-   */
-  private class InFlightMessage implements InvokeFuture<byte[]> {
-    private final NetworkVoltronEntityMessage message;
-    /**
-     * The set of pending ACKs determines when the caller returns from the send, in order to preserve ordering in the
-     * client code.  This is different from being "done" which specifically means that the APPLIED has happened,
-     * potentially returning a value or exception.
-     * ACKs are removed from this pending set, as they arrive.
-     */
-    private final Set<VoltronEntityMessage.Acks> pendingAcks;
-
-    private boolean isSent;
-    private EntityException exception;
-    private byte[] value;
-    private boolean done;
-
-    private InFlightMessage(NetworkVoltronEntityMessage message, Set<VoltronEntityMessage.Acks> acks) {
-      this.message = message;
-      this.pendingAcks = EnumSet.noneOf(VoltronEntityMessage.Acks.class);
-      this.pendingAcks.addAll(acks);
-    }
-
-    /**
-     * Used when populating the reconnect handshake.
-     */
-    public NetworkVoltronEntityMessage getMessage() {
-      return this.message;
-    }
-
-    public TransactionID getTransactionID() {
-      return this.message.getTransactionID();
-    }
-
-    public void send() {
-      Assert.assertFalse(this.isSent);
-      this.isSent = true;
-      this.message.send();
-    }
-
-    public synchronized void waitForAcks() {
-      boolean interrupted = false;
-      while (!this.pendingAcks.isEmpty()) {
-        try {
-          wait();
-        } catch (InterruptedException e) {
-          interrupted = true;
-        }
-      }
-      if (interrupted) {
-        Thread.currentThread().interrupt();
-      }
-    }
-
-    public synchronized void sent() {
-      if (this.pendingAcks.remove(VoltronEntityMessage.Acks.SENT)) {
-        if (this.pendingAcks.isEmpty()) {
-          notifyAll();
-        }
-      }
-    }
-
-    public synchronized void received() {
-      if (this.pendingAcks.remove(VoltronEntityMessage.Acks.RECEIVED)) {
-        if (this.pendingAcks.isEmpty()) {
-          notifyAll();
-        }
-      }
-    }
-
-    @Override
-    public void interrupt() {
-      throw new UnsupportedOperationException("Implement me!");
-    }
-
-    @Override
-    public synchronized boolean isDone() {
-      return done;
-    }
-
-    @Override
-    public synchronized byte[] get() throws InterruptedException, EntityException {
-      while (!done) {
-        wait();
-      }
-      if (exception != null) {
-        throw this.exception;
-      } else {
-        return value;
-      }
-    }
-
-    @Override
-    public synchronized byte[] getWithTimeout(long timeout, TimeUnit unit) throws InterruptedException, EntityException, TimeoutException {
-      long end = System.nanoTime() + unit.toNanos(timeout);
-      while (!done) {
-        long timing = end - System.nanoTime();
-        if (timing <= 0) {
-          throw new TimeoutException();
-        } else {
-          wait(timing / TimeUnit.MILLISECONDS.toNanos(1), (int)(timing % TimeUnit.MILLISECONDS.toNanos(1))); 
-        }
-      }
-      if (exception != null) {
-        throw this.exception;
-      } else {
-        return value;
-      }
-    }
-
-    synchronized void setResult(byte[] value, EntityException e) {
-      this.pendingAcks.remove(VoltronEntityMessage.Acks.APPLIED);
-      this.exception = e;
-      this.value = value;
-      this.done = true;
-      notifyAll();
-    }
   }
 }
